LPS-204779 Move UpgradeProcess to reuse it, but only run it once for bigger upgrades
