+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/internal/upgrade/v6_1_1/JournalArticleLayoutClassedModelUsageUpgradeProcess.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.journal.internal.upgrade.v4_4_3;
+package com.liferay.journal.internal.upgrade.v6_1_1;
 
 import com.liferay.asset.publisher.constants.AssetPublisherPortletKeys;
 import com.liferay.journal.model.JournalArticle;
