+++ modules/apps/journal/journal-test/src/testIntegration/java/com/liferay/journal/internal/upgrade/v6_1_1/test/JournalArticleLayoutClassedModelUsageUpgradeProcessTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.journal.internal.upgrade.v4_4_3.test;
+package com.liferay.journal.internal.upgrade.v6_1_1.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.asset.kernel.model.AssetEntry;
@@ -380,7 +380,7 @@ public class JournalArticleLayoutClassedModelUsageUpgradeProcessTest {
 	}
 
 	private static final String _CLASS_NAME =
-		"com.liferay.journal.internal.upgrade.v4_4_3." +
+		"com.liferay.journal.internal.upgrade.v6_1_1." +
 			"JournalArticleLayoutClassedModelUsageUpgradeProcess";
 
 	@Inject(
